scripts/margin_edge_integration/requirements.txt
python-dotenv
requests
PyMySQL
openpyxl